requirements.txt
numpy==1.18.1
scipy==1.4.1
matplotlib==3.1.3
tensorflow==1.15.2
opencv-python==4.2.0.34
opencv-contrib-python==4.2.0.34
tkinter==8.6
Pillow==5.3.0